fix: widen initial_supply to u128 in lib-client create_token builder
TokenCreationPayloadV1.initial_supply remains u64 (bincode compat) with
safe downcast at boundary. The public API accepts u128.

+++ lib-client/src/lib.rs
@@ -1457,7 +1457,7 @@ pub extern "C" fn zhtp_client_build_token_create(
     handle: *const IdentityHandle,
     name: *const std::ffi::c_char,
     symbol: *const std::ffi::c_char,
-    initial_supply: u64,
+    initial_supply: u128,
     decimals: u8,
     treasury_recipient: *const u8,
     chain_id: u8,
